requirements.txt
jupyter
ipykernel
pandas
numpy
holoviews
matplotlib
statsmodels
pyarrow
wfdb
websockets
